Move FIT data formats to DataFormats/Detectors, remove duplicate headers
