+++ DataFormats/Detectors/CMakeLists.txt
@@ -5,3 +5,4 @@ add_subdirectory (TPC)
 add_subdirectory (ITSMFT)
 add_subdirectory (MUON)
 add_subdirectory (TOF)
+add_subdirectory (FIT)
