+++ Detectors/FIT/T0/base/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(MODULE_NAME "T0Base")
+set(BUCKET_NAME fit_base_bucket)
 
 O2_SETUP(NAME ${MODULE_NAME})
 
@@ -12,7 +13,7 @@ set(HEADERS
 
   Set(LINKDEF src/T0BaseLinkDef.h)
 Set(LIBRARY_NAME ${MODULE_NAME})
-set(BUCKET_NAME fit_base_bucket)
+
 
 O2_GENERATE_LIBRARY()
