+++ Detectors/FIT/T0/reconstruction/src/T0ReconstructionLinkDef.h
@@ -15,9 +15,5 @@
 #pragma link off all functions;
 
 #pragma link C++ class o2::t0::CollisionTimeRecoTask + ;
-#pragma link C++ class o2::t0::RecPoints + ;
-#pragma link C++ class std::vector < o2::t0::RecPoints > +;
-
-#pragma link C++ class o2::t0::Channel + ;
 
 #endif
