+++ Detectors/FIT/common/CMakeLists.txt
@@ -1,4 +1,4 @@
 # Libraries
-add_subdirectory(base)
+#add_subdirectory(base)
 add_subdirectory(simulation)
 #add_subdirectory(reconstruction)
