+++ Detectors/FIT/common/simulation/include/FITSimulation/Digitizer.h
@@ -11,11 +11,12 @@
 #ifndef ALICEO2_FIT_DIGITIZER_H
 #define ALICEO2_FIT_DIGITIZER_H
 
-#include "FITBase/Digit.h"
+#include "CommonDataFormat/InteractionRecord.h"
+#include "DataFormatsFIT/Digit.h"
+#include "DataFormatsFIT/MCLabel.h"
 #include "T0Simulation/Detector.h"
 #include "SimulationDataFormat/MCTruthContainer.h"
 #include "SimulationDataFormat/MCCompLabel.h"
-#include "FITBase/MCLabel.h"
 #include "FITSimulation/DigitizationParameters.h"
 
 namespace o2
@@ -37,8 +38,16 @@ class Digitizer
   void setEventTime(double value) { mEventTime = value; }
   void setEventID(Int_t id) { mEventID = id; }
   void setSrcID(Int_t id) { mSrcID = id; }
-  void setBC(Int_t bc) { mBC = bc; }
-  void setOrbit(Int_t orbit) { mOrbit = orbit; }
+  void setInteractionRecord(uint16_t bc, uint32_t orbit)
+  {
+    mIntRecord.bc = bc;
+    mIntRecord.orbit = orbit;
+  }
+  const o2::InteractionRecord& getInteractionRecord() const { return mIntRecord; }
+  o2::InteractionRecord& getInteractionRecord(o2::InteractionRecord& src) { return mIntRecord; }
+  void setInteractionRecord(const o2::InteractionRecord& src) { mIntRecord = src; }
+  uint32_t getOrbit() const { return mIntRecord.orbit; }
+  uint16_t getBC() const { return mIntRecord.bc; }
 
   void setTriggers(Digit* digit);
   void smearCFDtime(Digit* digit);
@@ -52,8 +61,7 @@ class Digitizer
   // digit info
   // parameters
   Int_t mMode;  //triggered or continuos
-  Int_t mBC;    // BC
-  Int_t mOrbit; //orbit
+  o2::InteractionRecord mIntRecord; // Interaction record (orbit, bc)
   Int_t mEventID;
   Int_t mSrcID;        // signal, background or QED
   Double_t mEventTime; // timestamp
