+++ Steer/DigitizerWorkflow/src/FITDigitWriterSpec.cxx
@@ -14,8 +14,8 @@
 #include "Framework/CallbackService.h"
 #include "Framework/ControlService.h"
 #include "Framework/Task.h"
-#include "FITBase/Digit.h"
-#include "FITSimulation/MCLabel.h"
+#include "DataFormatsFIT/Digit.h"
+#include "DataFormatsFIT/MCLabel.h"
 #include "Headers/DataHeader.h"
 #include "DetectorsCommonDataFormats/DetID.h"
 #include "SimulationDataFormat/MCTruthContainer.h"
